move tulu3 tests to test_utils.py, clean up docstrings
Move test_tulu3_leading_newline_assistant_no_crash and
test_tulu3_leading_newline_via_chat_example from test_sft_tokenization.py
to tests/train/generators/test_utils.py where the tested function lives.

Rewrite _find_generation_prompt_boundary, _tokenize_chat_last_assistant,
and _tokenize_chat_all_assistants docstrings as proper function docs (what
it takes / what it returns); remove leftover PR-context notes and debug
comments from the function bodies.
