+++ tests/train/generators/test_utils.py
@@ -7,12 +7,14 @@ import os
 import pytest
 from transformers import AutoTokenizer
 
+from skyrl.train.config.sft_config import TrainOnWhat
 from skyrl.train.generators.utils import (
     apply_overlong_filtering,
     encode_messages_subset,
     get_generation_prompt_ids,
     get_response_ids_and_loss_mask_from_messages,
 )
+from skyrl.train.sft_trainer import tokenize_chat_example
 
 # Path to the custom Qwen3 chat template that doesn't add empty thinking blocks
 QWEN3_ACC_THINKING_TEMPLATE_PATH = os.path.join(
@@ -931,6 +933,90 @@ class TestGetResponseIdsAndLossMaskFromMessages:
 
         assert len(expected_loss_mask) == 32, "Total should be 32 tokens"
         assert loss_mask == expected_loss_mask, f"Expected {expected_loss_mask}, got {loss_mask}"
+
+
+# ============================================================================
+# Regression: TULU3-style assistant content starting with whitespace
+# ============================================================================
+
+
+@pytest.fixture(scope="module")
+def qwen25_tokenizer():
+    """Qwen2.5 tokenizer used to repro the TULU3 leading-newline merge case."""
+    tok = AutoTokenizer.from_pretrained("Qwen/Qwen2.5-0.5B-Instruct")
+    if tok.pad_token is None:
+        tok.pad_token = tok.eos_token
+    return tok
+
+
+def test_tulu3_leading_newline_assistant_no_crash(qwen25_tokenizer):
+    """Assistant content starting with ``\\n`` must not raise.
+
+    With Qwen2.5 and content ``"\\nHello"``, the header's trailing ``\\n``
+    (id 198) merges with the content's leading ``\\n`` into a single ``\\n\\n``
+    token (id 271) during tokenization.  ``_find_generation_prompt_boundary``
+    detects this and returns the pre-merge boundary index so the merged token
+    gets loss 1 (part of the assistant's generated content).
+    """
+    messages = [
+        {"role": "user", "content": "Hi"},
+        {"role": "assistant", "content": "\nHello world"},
+    ]
+
+    # Should not raise
+    response_ids, loss_mask, _ = get_response_ids_and_loss_mask_from_messages(messages, qwen25_tokenizer)
+
+    assert len(response_ids) == len(loss_mask)
+    assert qwen25_tokenizer.eos_token_id in response_ids
+
+    # The merged '\n\n' boundary token belongs to the assistant's generation
+    # window (loss 1). It starts right after the '<|im_start|>assistant'
+    # prefix which has length len(generation_prompt_ids) - 1 in the merge case.
+    gen_prompt = get_generation_prompt_ids(qwen25_tokenizer)
+
+    # Find indices of '<|im_start|>' (id 151644) — should be 2 (user, assistant).
+    im_start_id = 151644
+    im_start_indices = [i for i, t in enumerate(response_ids) if t == im_start_id]
+    assert len(im_start_indices) == 2, f"Expected two <|im_start|> tokens, got {len(im_start_indices)}"
+    assistant_turn_start = im_start_indices[1]
+
+    # First token of assistant turn is <|im_start|> (gen_prompt[0]), loss 0
+    assert loss_mask[assistant_turn_start] == 0
+    # Second token is 'assistant' (gen_prompt[1]), loss 0
+    assert loss_mask[assistant_turn_start + 1] == 0
+    # Third token is the merged '\n\n' (id 271) — loss 1 under our fix
+    assert response_ids[assistant_turn_start + 2] == 271, (
+        f"Expected merged '\\n\\n' token (271), got "
+        f"{response_ids[assistant_turn_start + 2]}. gen_prompt={gen_prompt}"
+    )
+    assert loss_mask[assistant_turn_start + 2] == 1
+
+    # At least some tokens must have loss=1 (the actual reply content)
+    assert sum(loss_mask) > 0
+
+
+def test_tulu3_leading_newline_via_chat_example(qwen25_tokenizer):
+    """End-to-end: ``tokenize_chat_example`` with ``ALL_ASSISTANT_MESSAGES`` must
+    not crash when an assistant message starts with ``\\n``."""
+    example = {
+        "messages": [
+            {"role": "system", "content": "You are helpful."},
+            {"role": "user", "content": "Give me a poem."},
+            {"role": "assistant", "content": "\nRoses are red,\nViolets are blue."},
+        ]
+    }
+
+    result = tokenize_chat_example(
+        example,
+        qwen25_tokenizer,
+        train_on_what=TrainOnWhat.ALL_ASSISTANT_MESSAGES,
+    )
+
+    assert result is not None
+    assert "loss_mask" in result
+    assert len(result["loss_mask"]) == result["num_actions"]
+    assert sum(result["loss_mask"]) > 0
+    assert all(v in (0, 1) for v in result["loss_mask"])
 
 
 # ============================================================================
